% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optObj_sybilMosekClass.R
\name{delProb,optObj_sybilMosek-method}
\alias{delProb,optObj_sybilMosek-method}
\title{An S4 class to remove a optObj class.}
\usage{
\S4method{delProb}{optObj_sybilMosek}(lp, ...)
}
\description{
An S4 class to remove a optObj class.
}
