% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optObj_sybilMosekClass.R
\name{initProb,optObj_sybilMosek-method}
\alias{initProb,optObj_sybilMosek-method}
\title{An S4 class to initialize the optObj_sybilMosek class.}
\usage{
\S4method{initProb}{optObj_sybilMosek}(lp, to = FALSE, nrows = 0, ncols = 0, ...)
}
\description{
An S4 class to initialize the optObj_sybilMosek class.
}
\section{Slots}{

\describe{
\item{\code{lp}}{}
}}

