% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optObj_sybilMosekClass.R
\docType{class}
\name{optObj_sybilMosek-class}
\alias{optObj_sybilMosek-class}
\title{An S4 class to represent a optObj for SybilMosek}
\description{
An S4 class to represent a optObj for SybilMosek
}
